ROFO-198 refactor: bottomButton 컴포넌트화
